requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Flask==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0
google-auth==2.40.2
google-generativeai==0.3.0
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.62.3
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
Pygments==2.19.1
pytest==8.4.0
python-dotenv==1.0.0
python-espeak==0.6.3
pyttsx3==2.90
requests==2.31.0
rsa==4.9.1
SpeechRecognition==3.10.0
tqdm==4.67.1
urllib3==2.4.0
Werkzeug==3.1.3
